Created a separated module for the MLPModel.
