+++ beer/models/__init__.py
@@ -5,6 +5,7 @@ from .normal import NormalFullCovariance
 from .mixture import Mixture
 
 from .vae import VAE
-from .vae import MLPNormalDiag
-from .vae import MLPNormalIso
+
+from .mlpmodel import MLPNormalDiag
+from .mlpmodel import MLPNormalIso
