+++ beer/models/vae.py
@@ -1,8 +1,8 @@
 
-"""Implementation of the Variational Auto-Encoder with arbitrary
+'''Implementation of the Variational Auto-Encoder with arbitrary
 prior over the latent space.
 
-"""
+'''
 
 import abc
 import math
@@ -10,17 +10,13 @@ import math
 import torch
 from torch import nn
 from torch.autograd import Variable
-import numpy as np
-
-from .model import Model
-from ..training import mini_batches
 
 
 class VAE(nn.Module):
-    """Variational Auto-Encoder (VAE)."""
+    '''Variational Auto-Encoder (VAE).'''
 
     def __init__(self, encoder, decoder, latent_model, nsamples):
-        """Initialize the VAE.
+        '''Initialize the VAE.
 
         Args:
             encoder (``MLPModel``): Encoder of the VAE.
@@ -30,7 +26,7 @@ class VAE(nn.Module):
             nsamples (int): Number of samples to approximate the
                 expectation of the log-likelihood.
 
-        """
+        '''
         super().__init__()
         self.encoder = encoder
         self.decoder = decoder
@@ -91,7 +87,7 @@ class VAE(nn.Module):
         }
 
     def loss(self, X, state, kl_weight=1.0):
-        """Loss function of the VAE. This is the negative of the
+        '''Loss function of the VAE. This is the negative of the
         variational objective function i.e.:
 
             loss = - ( E_q [ ln p(X|Z) ] - KL( q(z) || p(z) ) )
@@ -107,7 +103,7 @@ class VAE(nn.Module):
         Returns:
             torch.Variable: Symbolic computation of the loss function.
 
-        """
+        '''
         nsamples = state['nsamples']
         llh = state['decoder_state'].log_likelihood(X, state['nsamples'])
         llh = llh.view(nsamples, X.size(0), -1).sum(dim=0) / nsamples
@@ -115,211 +111,4 @@ class VAE(nn.Module):
         kl *= kl_weight
 
         return -(llh - kl[:, None]), llh, kl
-
-
-class MLPEncoderState(metaclass=abc.ABCMeta):
-    'Abstract Base Class for the state of a the VAE encoder.'
-
-    @property
-    def mean(self):
-        'Mean of each distribution.'
-        return self._mean
-
-    @property
-    def prec(self):
-        'Diagonal of the precision matrix for each distribution.'
-        return self._prec
-
-    @abc.abstractmethod
-    def sample(self):
-        'sample data using the reparametization trick.'
-        NotImplemented
-
-    @abc.abstractmethod
-    def kl_div(self, p_nparams):
-        'kl divergence between the posterior and prior distribution.'
-        NotImplemented
-
-
-class MLPDecoderState(metaclass=abc.ABCMeta):
-    'Abstract Base Class for the state of a the VAE decoder.'
-
-    @abc.abstractmethod
-    def natural_params(self):
-        'Natural parameters for each distribution.'
-        NotImplemented
-
-    @abc.abstractmethod
-    def log_base_measure(self, X):
-        'Natural parameters for each distribution.'
-        NotImplemented
-
-    @abc.abstractmethod
-    def sufficient_statistics(self, X):
-        'Sufficient statistics of the given data.'
-        NotImplemented
-
-    def log_likelihood(self, X, nsamples=1):
-        'Log-likelihood of the data.'
-        s_stats = self.sufficient_statistics(X)
-        nparams = self.natural_params()
-        log_bmeasure = self.log_base_measure(X)
-        nparams = nparams.view(nsamples, X.size(0), -1)
-        return torch.sum(nparams * s_stats, dim=-1) + log_bmeasure
-
-
-class MLPStateNormal(MLPEncoderState, MLPDecoderState):
-
-    def __init__(self, mean, prec):
-        self._mean = mean
-        self._prec = prec
-
-    def exp_T(self):
-        idxs = torch.arange(0, self.mean.size(1)).long()
-        XX = self.mean[:, :, None] * self.mean[:, None, :]
-        XX[:, idxs, idxs] += 1 / self.prec
-        return torch.cat([XX.view(self.mean.size(0), -1), self.mean,
-                          Variable(torch.ones(self.mean.size(0), 2))], dim=-1)
-
-    @property
-    def std_dev(self):
-        return 1 / torch.sqrt(self.prec)
-
-    def natural_params(self):
-        identity = Variable(torch.eye(self.mean.size(1)))
-        np1 = -.5 * self.prec[:, None] * identity[None, :, :]
-        np1 = np1.view(self.mean.size(0), -1)
-        np2 = self.prec * self.mean
-        np3 = -.5 * (self.prec * (self.mean ** 2)).sum(-1)[:, None]
-        np4 = .5 * torch.log(self.prec).sum(-1)[:, None]
-        return torch.cat([np1, np2, np3, np4], dim=-1)
-
-    def sample(self):
-        noise = Variable(torch.randn(*self.mean.size()))
-        return self.mean + self.std_dev * noise
-
-    def kl_div(self, p_nparams):
-        return ((self.natural_params() - p_nparams) * self.exp_T()).sum(dim=-1)
-
-    def sufficient_statistics(self, X):
-        XX = X[:, :, None] * X[:, None, :]
-        return torch.cat([XX.view(X.size(0), -1), X,
-                          Variable(torch.ones(X.size(0), 2).float())], dim=-1)
-
-    def log_base_measure(self, X):
-        return -.5 * X.size(-1) * math.log(2 * math.pi)
-
-
-class MLPModel(nn.Module):
-    '''Base class for the encoder / decoder neural network of
-    the VAE. The output of this network are the parameters of a
-    conjugate exponential model. The proper way to use this class
-    is to wrap with an object that "knows" how to make sense of the
-    outptuts (see ``MLPEncoderState``, ``MLPDecoderIso``, ...).
-
-    Note:
-        This class only defines the neural network structure and does
-        not care wether it is used as encoder/decoder and how the
-        parameters of the model is used.
-
-    '''
-
-    @staticmethod
-    def _init_residulal_layer(linear_transform):
-        W = linear_transform.weight.data.numpy()
-        dim = max(*W.shape)
-        q, _ = np.linalg.qr(np.random.randn(dim, dim))
-        W = q[:W.shape[0], :W.shape[1]]
-        linear_transform.weight = nn.Parameter(torch.from_numpy(W).float())
-
-    def __init__(self, structure, outputs):
-        '''Initialize the ``MLPModel``.
-
-        Args:
-            structure (``torch.Sequential``): Sequence linear/
-                non-linear operations.
-            outputs (list): List of tuple describing the output model.
-
-        '''
-        super().__init__()
-        self.structure = structure
-
-        # Get the input/ouput dimension of the structure.
-        for transform in structure:
-            if isinstance(transform, nn.Linear):
-                in_dim = transform.in_features
-                break
-        for transform in reversed(structure):
-            if isinstance(transform, nn.Linear):
-                out_dim = transform.out_features
-                break
-
-        # Create the specific output layer.
-        self.output_layer = nn.ModuleList()
-        self.residual_connections = nn.ModuleList()
-        self.residual_mapping = {}
-        for i, output in enumerate(outputs):
-            target_dim, residual = output
-            self.output_layer.append(nn.Linear(out_dim, target_dim))
-            if residual:
-                ltransform = nn.Linear(in_dim, target_dim)
-                MLPModel._init_residulal_layer(ltransform)
-                self.residual_connections.append(ltransform)
-                self.residual_mapping[i] = len(self.residual_connections) - 1
-
-    def forward(self, X):
-        h = self.structure(X)
-        outputs = [transform(h) for transform in self.output_layer]
-        for idx1, idx2 in self.residual_mapping.items():
-            outputs[idx1] += self.residual_connections[idx2](X)
-        return outputs
-
-class MLPNormalDiag(MLPModel):
-    '''Neural-Network ending with a double linear projection
-    providing the mean and the logarithm of the diagonal of the
-    covariance matrix.
-
-    '''
-
-    def __init__(self, structure, dim, residual=False):
-        '''Initialize a ``MLPNormalDiag`` object.
-
-        Args:
-            structure (``torch.Sequential``): Sequence linear/
-                non-linear operations.
-            dim (int): Desired dimension of the modeled random
-                variable.
-            residual (boolean): Add a residual connection to the mean.
-
-        '''
-        super().__init__(structure, [(dim, residual), (dim, False)])
-
-    def forward(self, X):
-        mean, logprec = super().forward(X)
-        return MLPStateNormal(mean, torch.exp(logprec))
-
-
-class MLPNormalIso(MLPModel):
-    '''Neural-Network ending with a double linear projection
-    providing the mean and the isotropic covariance matrix.
-
-    '''
-
-    def __init__(self, structure, dim, residual=False):
-        '''Initialize a ``MLPNormalDiag`` object.
-
-        Args:
-            structure (``torch.Sequential``): Sequence linear/
-                non-linear operations.
-            dim (int): Desired dimension of the modeled random
-                variable.
-            residual (boolean): Add a residual connection to the mean.
-
-        '''
-        super().__init__(structure, [(dim, residual), (1, False)])
-
-    def forward(self, X):
-        mean, logprec = super().forward(X)
-        return MLPStateNormal(mean,
-            torch.exp(logprec) * Variable(torch.ones(mean.size(1)).float()))
